requirements.txt
discord.py>=1.4.1
num2words>=0.5.10